Play around with ncurses
I've built an abstraction over ncurses to make "rendering" a little bit easier. I can write characters (sprites) to a grid data structure (window surface) and write the characters to the terminal (swap buffers). I'm going to try to lay out the code in the same way I would if this was a regular OpenGL SDL2 application.