Scrap curses
Curses is not as ubiquitous as I thought it was. I thought that using curses will make building easie
r but I was wrong. Curses is not everywhere. Not all terminals support 256 color modes. ASCII art loo
ks kind of ugly.

I decided to use a library that I have been using for years: SDL2. Installing SDL2 is pretty easy. You just download the sources and run make. On a mac, you just `brew install sdl2`.

I won't use OpenGL though. Getting OpenGL working everywhere is going to be super annoying. The SDL_Render API is unpleasant to use but it's OK for a really simple game like this one.