Handle potential hang
The setScaredTargetSystem has the potential to hang if a scared ghost reaches a dead end. This can only happen when inside the ghost house. I plan to add EnterHouse and LeaveHouse components and systems to handle entering and leaving the ghost house so that ghosts can never be scared and find themselves in a dead end.

So this fix is kind of pointless really!